+++ packages/menu/src/mwc-menu-base.ts
@@ -17,21 +17,21 @@ limitations under the License.
 import '@material/mwc-list';
 import './mwc-menu-surface';
 
-import {MDCMenuAdapter} from '@material/menu/adapter.js';
-import {DefaultFocusState as DefaultFocusStateEnum} from '@material/menu/constants.js';
-import MDCMenuFoundation from '@material/menu/foundation.js';
-import {BaseElement} from '@material/mwc-base/base-element.js';
-import {observer} from '@material/mwc-base/observer.js';
+import {MDCMenuAdapter} from '@material/menu/adapter';
+import {DefaultFocusState as DefaultFocusStateEnum} from '@material/menu/constants';
+import MDCMenuFoundation from '@material/menu/foundation';
+import {BaseElement} from '@material/mwc-base/base-element';
+import {observer} from '@material/mwc-base/observer';
 import {List, MWCListIndex} from '@material/mwc-list';
-import {ActionDetail} from '@material/mwc-list/mwc-list-foundation.js';
-import {ListItemBase} from '@material/mwc-list/mwc-list-item-base.js';
+import {ActionDetail} from '@material/mwc-list/mwc-list-foundation';
+import {ListItemBase} from '@material/mwc-list/mwc-list-item-base';
 import {html, property, query} from 'lit-element';
 
 import {MenuSurface} from './mwc-menu-surface';
-import {Corner} from './mwc-menu-surface-base';
+import {Corner, MenuCorner} from './mwc-menu-surface-base';
 
 export {createSetFromIndex, isEventMulti, isIndexSet, MWCListIndex} from '@material/mwc-list/mwc-list-foundation';
-export {Corner} from './mwc-menu-surface-base';
+export {Corner, MenuCorner} from './mwc-menu-surface-base';
 
 export type DefaultFocusState = keyof typeof DefaultFocusStateEnum;
 
@@ -79,6 +79,8 @@ export abstract class MenuBase extends BaseElement {
   @property({type: Boolean}) forceGroupSelection = false;
 
   @property({type: Boolean}) fullwidth = false;
+
+  @property({type: String}) menuCorner: MenuCorner = 'START';
 
   @property({type: String})
   @observer(function(this: MenuBase, value: DefaultFocusState) {
@@ -144,6 +146,7 @@ export abstract class MenuBase extends BaseElement {
           .absolute=${this.absolute}
           .fixed=${this.fixed}
           .fullwidth=${this.fullwidth}
+          .menuCorner=${this.menuCorner}
           class="mdc-menu mdc-menu-surface"
           @closed=${this.onClosed}
           @opened=${this.onOpened}