+++ packages/menu/src/mwc-menu.ts
@@ -22,7 +22,7 @@ import {style} from './mwc-menu-css.js';
 
 export {DefaultFocusState} from '@material/menu/constants.js';
 export {createSetFromIndex, isEventMulti, isIndexSet, MWCListIndex} from '@material/mwc-list/mwc-list-foundation.js';
-export {Corner} from './mwc-menu-surface-base.js';
+export {Corner, MenuCorner} from './mwc-menu-surface-base.js';
 
 declare global {
   interface HTMLElementTagNameMap {